@@ -1,10 +1,7 @@
 import { Component, EventEmitter, Input, OnInit, Output } from '@angular/core';
-import { CommonModule } from '@angular/common';
 import { Mail } from 'src/app/model/mail';
-import { MaterialModule } from 'src/app/material-module/material/material.module';
 import { FolderService } from 'src/app/services/folder.service';
-import { ActivatedRoute, Router, RouterStateSnapshot } from '@angular/router';
-import { MessageActionsComponent } from '../message-actions/message-actions.component';
+import { ActivatedRoute, Router } from '@angular/router';
 import { ContactsService } from 'src/app/services/contacts.service';
 import { MatSnackBar } from '@angular/material/snack-bar';
 import { Contact } from 'src/app/model/contact';
@@ -31,7 +28,7 @@ export class MessageViewerComponent implements OnInit {
   ) {}
 
   ngOnInit() {
-    if(this.selectedMessage){const mailId = this.route.snapshot.params['id'];
+    const mailId = this.route.snapshot.params['id'];
     const folderName = this.route.snapshot.params['folderName'];
   
     this.folderService.getEmailsObservable(folderName).forEach(
@@ -50,11 +47,6 @@ export class MessageViewerComponent implements OnInit {
      
     )}
     
-  }
-  
-  
-  
-  
   
 
   replyToEmail() {
